pyproject.toml
[project]
name = "nocost-backend"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "pydantic[email]",
    "python-multipart",
    "ansible",
    "alembic",
    "aiomysql",
    "dnspython",
    "sqlalchemy[asyncio]",
    "bcrypt",
    "httpx",
    "PyJWT",
    "uvicorn",
    "gunicorn",
    "ansible-runner",
    "python-dotenv",
    "geoip2",
]

[project.optional-dependencies]
dev = [
    "httpx",
    "aiosqlite",
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
backup_site = "cli.backup_site:main"
backup_system = "cli.backup_system:main"
create_site = "cli.create_site:main"
link_domain = "cli.link_domain:main"
list_sites = "cli.list_sites:main"
remove_site = "cli.remove_site:main"
restore_site = "cli.restore_site:main"
make_donor = "cli.make_donor:main"
site_info = "cli.site_info:main"
sync_nginx_maps = "cli.sync_nginx_maps:main"
upgrade_site = "cli.upgrade_site:main"

[tool.pytest.ini_options]
testpaths = ["tests/unit"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
